Add files via upload
#1 project
Hangman game is just a word guessing game by guess the character of the word. In this game, there is a list of words present, out of which our interpreter will choose 1 random word.

The user first has to input their names and then, will be asked to guess any alphabet. If the random word contains that alphabet, it will be shown as the output(with correct placement) else the program will ask you to guess another alphabet. Users will be given 10 turns(can be changed accordingly) to guess the complete word.

TODO
Change the turns that you want to give the player
Change the list containing Secret word list